% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{FutureManager demo app}
\usage{
demo(host = "0.0.0.0", port = 3838L, launch.browser = TRUE)
}
\arguments{
\item{host}{host IP}

\item{port}{app port}

\item{launch.browser}{logical, should launch the browser?}
}
\description{
Runs the demo app that shows the abilities of the package
}
