% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_noButton}
\alias{demo_noButton}
\title{FutureManager demo app without the run button}
\usage{
demo_noButton(host = "0.0.0.0", port = 3838L, launch.browser = FALSE)
}
\arguments{
\item{host}{host IP}

\item{port}{app port}

\item{launch.browser}{logical, should launch the browser?}
}
\description{
Runs another demo app, that doesn't use run button (heavy calculations start 
on-demand, without additional user action)
}
