% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmError}
\alias{fmError}
\title{Error class}
\usage{
fmError(msg)
}
\arguments{
\item{msg}{Message that should be displayed to the user}
}
\value{
fmError object
}
\description{
Allows to signalize some error in the long function (not enough data to run 
calculations etc)
}
