% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmGenerateTaskId}
\alias{fmGenerateTaskId}
\title{Generate task ID}
\usage{
fmGenerateTaskId(id)
}
\arguments{
\item{id}{character string, ID pattern}
}
\value{
character string
}
\description{
Use this function if when you don't need an fmRunButton.
}
\details{
See \code{\link{demo_noButton}} for some example
}
