% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmIsInterrupted}
\alias{fmIsInterrupted}
\title{Check if process has been interrupted}
\usage{
fmIsInterrupted(task)
}
\arguments{
\item{task}{task object}
}
\value{
logical
}
\description{
Basically, it checks if the process has been canceled by the user.
}
