% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmNeed}
\alias{fmNeed}
\title{Need a valid background process value}
\usage{
fmNeed(x, msgInit = "run the process first", msgRun = "wait for the process")
}
\arguments{
\item{x}{fmStatus object}

\item{msgInit}{character string, message that should be displayed when the 
process is not run yet}

\item{msgRun}{character string, message that should be displayed when the
process is still running}
}
\value{
list of shiny needs
}
\description{
Ensures the value is valid.
}
