% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{fmRunButton}
\alias{fmRunButton}
\title{Run button}
\usage{
fmRunButton(inputId, fm, defaultValue = FALSE, blocked = FALSE)
}
\arguments{
\item{inputId}{character string, the button ID}

\item{fm}{FutureManager object}

\item{defaultValue}{logical, the initial button value}

\item{blocked}{logical, should the button be blocked?}
}
\value{
button HTML
}
\description{
Creates toggle button, that may be used to control a long background process (start, cancel etc)
}
\details{
See \code{\link{fmError}} for some example.
}
