% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmStatus}
\alias{fmStatus}
\title{Status class}
\usage{
fmStatus(id, status, message, value = NULL)
}
\arguments{
\item{id}{character string, task ID}

\item{status}{character string}

\item{message}{message that may be used on front-end to inform user what's happeninged}

\item{value}{NULL for not-"success" tasks, the long function value for "success" tasks}
}
\value{
fmStatus object
}
\description{
This class is used to pass the background process results and metadata to 
the app. You can access the value directly (i.e. obj$value) or via 
fmGetValue() function (recommended!).
}
\details{
Possible statuses:
- "running": task is running;
- "failed": task has been stopped using fmError(); 
- "success": task completed succesfully;
- "error": unexpected error occured in the long function;
- "canceled": task has been canceled by the user

This function should NOT be used directly.
}
\examples{
fmStatus("task1", "success", "Task completed successfully", iris)
}
