% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmUpdateProgress}
\alias{fmUpdateProgress}
\title{Update progress of the background process}
\usage{
fmUpdateProgress(task, progress = 0, msg = NULL)
}
\arguments{
\item{task}{task object}

\item{progress}{numeric in range 0..1}

\item{msg}{character string, message that will be displayed in the progress bar}
}
\value{
nothing
}
\description{
This function allows to pass progress info from the background process
to the main process, i.e. to display a progress in the application
}
