% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{fmUpdateRunButton}
\alias{fmUpdateRunButton}
\title{Update run button}
\usage{
fmUpdateRunButton(
  inputId,
  status,
  fm,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{character string, the button ID}

\item{status}{character string, the process status}

\item{fm}{FutureManager object}

\item{session}{shiny session object}
}
\value{
No return value, called for side effects.
}
\description{
Updates the run button (on frontend) and also its state (on backend)
}
