% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmValidate}
\alias{fmValidate}
\title{Validate the process value}
\usage{
fmValidate(x, ...)
}
\arguments{
\item{x}{fmStatus object}

\item{...}{arguments passed to fmNeed function (msgInit and msgRun)}
}
\value{
shiny validation
}
\description{
This function should be used in every shiny output, which uses the background 
process value directly. It ensures that the value is valid and displays an
error message in case of any issues. Also, it displays helpful messages regarding 
the process start/wait with a CSS formatting.
}
\details{
See \code{\link{fmError}} for some example.
}
