% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.fmError}
\alias{is.fmError}
\title{Check if object is of fmError class}
\usage{
is.fmError(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Check if object is of fmError class
}
