% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{FutureManager demo app}
\usage{
demo(host = "0.0.0.0", port = 3838L, launch.browser = FALSE)
}
\arguments{
\item{host}{host IP}

\item{port}{app port}

\item{launch.browser}{logical, should launch the browser?}
}
\value{
No return value, called for side effects.
}
\description{
Runs the demo app that shows the abilities of the package
}
\details{
Demo files location: \code{system.file("demoapp", package = "FutureManager")}
}
\examples{
if(interactive()) {
demo(launch.browser = TRUE)
}
}
