% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plan}
\alias{sequential}
\alias{transparent}
\alias{multisession}
\alias{multicore}
\alias{multiprocess}
\alias{cluster}
\alias{remote}
\alias{supportsMulticore}
\alias{\%>\%}
\alias{arrange}
\alias{mutate}
\alias{filter}
\alias{select}
\alias{rename}
\alias{tbl_vars}
\alias{group_vars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{tbl_vars}}}

  \item{future}{\code{\link[future]{cluster}}, \code{\link[future]{multicore}}, \code{\link[future]{multiprocess}}, \code{\link[future]{multisession}}, \code{\link[future]{plan}}, \code{\link[future]{remote}}, \code{\link[future]{sequential}}, \code{\link[future:re-exports]{supportsMulticore}}, \code{\link[future:sequential]{transparent}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

